refactor(svelte): split keyboard pure unit tests by decision group
Move shared SurfaceKeyboardInput builders to keyboard-fixtures and split
the 443-line keyboard.test.ts into area, inspect/escape, and misc suites.
Assertions preserved.
